move authorization validation functions
