+++ hack/import-restrictions.json
@@ -209,8 +209,7 @@
       "vendor/k8s.io/kubernetes/pkg/apis/core/v1",
       "vendor/k8s.io/kubernetes/pkg/apis/rbac",
       "vendor/k8s.io/kubernetes/pkg/apis/rbac/v1",
-      "vendor/k8s.io/kubernetes/pkg/api/legacyscheme",
-      "github.com/openshift/origin/pkg/authorization/apis/authorization/internal/serviceaccount"
+      "vendor/k8s.io/kubernetes/pkg/api/legacyscheme"
     ]
   },
